+++ QueryString.php
@@ -13,29 +13,15 @@ declare(strict_types=1);
 
 namespace League\Uri;
 
-use League\Uri\Contracts\UriComponentInterface;
 use League\Uri\Exceptions\SyntaxError;
+use League\Uri\KeyValuePair\Converter;
 use Stringable;
 use function array_key_exists;
 use function array_keys;
-use function explode;
-use function html_entity_decode;
-use function implode;
 use function is_array;
-use function is_bool;
-use function is_numeric;
-use function is_scalar;
-use function is_string;
-use function preg_match;
-use function preg_quote;
-use function preg_replace_callback;
 use function rawurldecode;
-use function rawurlencode;
-use function sprintf;
-use function str_replace;
 use function strpos;
 use function substr;
-use const PHP_QUERY_RFC1738;
 use const PHP_QUERY_RFC3986;
 
 /**
@@ -45,30 +31,83 @@ use const PHP_QUERY_RFC3986;
  */
 final class QueryString
 {
-    private const ENCODING_LIST = [
-        PHP_QUERY_RFC1738 => [
-            'suffixKey' => '*',
-            'suffixValue' => '*&',
-        ],
-        PHP_QUERY_RFC3986 => [
-            'suffixKey' => "!$'()*,;:@?/%",
-            'suffixValue' => "!$'()*,;=:@?/&%",
-        ],
-    ];
     private const PAIR_VALUE_DECODED = 1;
     private const PAIR_VALUE_PRESERVED = 2;
-    private const REGEXP_ENCODED_PATTERN = ',%[A-Fa-f0-9]{2},';
-    private const REGEXP_INVALID_CHARS = '/[\x00-\x1f\x7f]/';
-    private const REGEXP_UNRESERVED_CHAR = '/[^A-Za-z0-9_\-.~]/';
-
-    private static string $regexpKey;
-    private static string $regexpValue;
 
     /**
      * @codeCoverageIgnore
      */
     private function __construct()
     {
+    }
+
+    /**
+     * Build a query string from a list of pairs.
+     *
+     * @see QueryString::buildFromPairs()
+     * @see https://datatracker.ietf.org/doc/html/rfc3986#section-2.2
+     *
+     * @param iterable<array{0:string, 1:string|float|int|bool|null}> $pairs
+     * @param non-empty-string                                        $separator
+     *
+     * @throws SyntaxError If the encoding type is invalid
+     * @throws SyntaxError If a pair is invalid
+     */
+    public static function build(iterable $pairs, string $separator = '&', int $encType = PHP_QUERY_RFC3986): ?string
+    {
+        return self::buildFromPairs($pairs, Converter::fromEncodingType($encType)->withSeparator($separator));
+    }
+
+    /**
+     * Build a query string from a list of pairs.
+     *
+     * The method expects the return value from Query::parse to build
+     * a valid query string. This method differs from PHP http_build_query as
+     * it does not modify parameters keys.
+     *
+     *  If a reserved character is found in a URI component and
+     *  no delimiting role is known for that character, then it must be
+     *  interpreted as representing the data octet corresponding to that
+     *  character's encoding in US-ASCII.
+     *
+     * @see https://datatracker.ietf.org/doc/html/rfc3986#section-2.2
+     *
+     * @param iterable<array{0:string, 1:string|float|int|bool|null}> $pairs
+     *
+     * @throws SyntaxError If the encoding type is invalid
+     * @throws SyntaxError If a pair is invalid
+     */
+    public static function buildFromPairs(iterable $pairs, Converter $converter = null): ?string
+    {
+        $keyValuePairs = [];
+        foreach ($pairs as $pair) {
+            if ([0, 1] !== array_keys($pair)) { /* @phpstan-ignore-line */
+                throw new SyntaxError('A pair must be a sequential array starting at `0` and containing two elements.');
+            }
+
+            $keyValuePairs[] = [(string) Encoder::encodeQueryKeyValue($pair[0]), match(true) {
+                null === $pair[1] => null,
+                default => Encoder::encodeQueryKeyValue($pair[1]),
+            }];
+        }
+
+        return ($converter ?? Converter::fromRFC3986())->toValue($keyValuePairs);
+    }
+
+    /**
+     * Parses the query string like parse_str without mangling the results.
+     *
+     * @see QueryString::extractFromValue()
+     * @see http://php.net/parse_str
+     * @see https://wiki.php.net/rfc/on_demand_name_mangling
+     *
+     * @param non-empty-string $separator
+     *
+     * @throws SyntaxError
+     */
+    public static function extract(Stringable|string|bool|null $query, string $separator = '&', int $encType = PHP_QUERY_RFC3986): array
+    {
+        return self::extractFromValue($query, Converter::fromEncodingType($encType)->withSeparator($separator));
     }
 
     /**
@@ -83,16 +122,26 @@ final class QueryString
      *
      * @throws SyntaxError
      */
-    public static function extract(Stringable|string|bool|null $query, string $separator = '&', int $encType = PHP_QUERY_RFC3986): array
+    public static function extractFromValue(Stringable|string|bool|null $query, Converter $converter = null): array
     {
-        $query = self::filterQuery($query, $encType);
+        return self::convert(self::decodePairs(
+            ($converter ?? Converter::fromRFC3986())->toPairs($query),
+            self::PAIR_VALUE_PRESERVED
+        ));
+    }
 
-        return self::convert(match (true) {
-            '' === $separator => throw new SyntaxError('The separator character can not be the empty string.'),
-            null === $query,
-            '' === $query => [],
-            default => self::parsePairs($query, $separator, self::PAIR_VALUE_PRESERVED),
-        });
+    /**
+     * Parses a query string into a collection of key/value pairs.
+     *
+     * @param non-empty-string $separator
+     *
+     * @throws SyntaxError
+     *
+     * @return array<int, array{0:string, 1:string|null}>
+     */
+    public static function parse(Stringable|string|bool|null $query, string $separator = '&', int $encType = PHP_QUERY_RFC3986): array
+    {
+        return self::parseFromValue($query, Converter::fromEncodingType($encType)->withSeparator($separator));
     }
 
     /**
@@ -102,81 +151,35 @@ final class QueryString
      *
      * @return array<int, array{0:string, 1:string|null}>
      */
-    public static function parse(Stringable|string|bool|null $query, string $separator = '&', int $encType = PHP_QUERY_RFC3986): array
+    public static function parseFromValue(Stringable|string|bool|null $query, Converter $converter = null): array
     {
-        $query = self::filterQuery($query, $encType);
-
-        return match (true) {
-            '' === $separator => throw new SyntaxError('The separator character can not be the empty string.'),
-            null === $query => [],
-            '' === $query => [['', null]],
-            default => self::parsePairs($query, $separator, self::PAIR_VALUE_DECODED),
-        };
-    }
-
-    private static function filterQuery(Stringable|string|bool|null $query, int $encType): ?string
-    {
-        $query = match (true) {
-            !isset(self::ENCODING_LIST[$encType]) => throw new SyntaxError('Unknown or Unsupported encoding.'),
-            $query instanceof UriComponentInterface => $query->value(),
-            $query instanceof Stringable => (string) $query,
-            default => $query,
-        };
-
-        return match (true) {
-            null === $query => null,
-            false === $query => '0',
-            true === $query => '1',
-            1 === preg_match(self::REGEXP_INVALID_CHARS, $query) => throw new SyntaxError(sprintf('Invalid query string: %s.', $query)),
-            PHP_QUERY_RFC1738 === $encType => str_replace('+', '%20', $query),
-            default => $query,
-        };
-    }
-
-    /**
-     * @param non-empty-string $query
-     * @param non-empty-string $separator
-     *
-     * @return array<int, array{0:string, 1:string|null}>
-     */
-    private static function parsePairs(string $query, string $separator, int $pairValueState): array
-    {
-        return array_reduce(
-            match (true) {
-                str_contains($query, $separator) => explode($separator, $query),
-                default => [$query],
-            },
-            fn (array $carry, string|null $pairString) => [...$carry, self::parsePair((string) $pairString, $pairValueState)],
-            []
+        return self::decodePairs(
+            ($converter ?? Converter::fromRFC3986())->toPairs($query),
+            self::PAIR_VALUE_DECODED
         );
     }
 
     /**
-     * Returns the key/value pair from a query string pair.
+     * @param array<non-empty-list<string|null>> $pairs
      *
-     * @return array{0:string, 1:string|null}
+     * @return array<int, array{0:string, 1:string|null}>
      */
-    private static function parsePair(string $pair, int $pairValueState): array
+    private static function decodePairs(array $pairs, int $pairValueState): array
     {
-        $decodeMatches = static fn (array $matches): string => rawurldecode($matches[0]);
+        $decodePair = static function (array $pair, int $pairValueState): array {
+            [$key, $value] = $pair;
 
-        [$key, $value] = explode('=', $pair, 2) + [1 => null];
-
-        $key = match (true) {
-            null === $key => '',
-            1 !== preg_match(self::REGEXP_ENCODED_PATTERN, $key) => $key,
-            default => (string) preg_replace_callback(self::REGEXP_ENCODED_PATTERN, $decodeMatches(...), $key),
+            return match (true) {
+                self::PAIR_VALUE_PRESERVED === $pairValueState => [(string) Encoder::decodeAll($key), $value],
+                default => [(string) Encoder::decodeAll($key), Encoder::decodeAll($value)],
+            };
         };
 
-        return match (true) {
-            null === $value,
-            self::PAIR_VALUE_PRESERVED === $pairValueState,
-            1 !== preg_match(self::REGEXP_ENCODED_PATTERN, $value) => [$key, $value],
-            default => [
-                $key,
-                preg_replace_callback(self::REGEXP_ENCODED_PATTERN, $decodeMatches(...), $value),
-            ],
-        };
+        return array_reduce(
+            $pairs,
+            fn (array $carry, array $pair) => [...$carry, $decodePair($pair, $pairValueState)],
+            []
+        );
     }
 
     /**
@@ -263,94 +266,5 @@ final class QueryString
         $data[$key] = self::extractPhpVariable($data[$key], $index.$remaining, $value);
 
         return $data;
-    }
-
-    /**
-     * Build a query string from an associative array.
-     *
-     * The method expects the return value from Query::parse to build
-     * a valid query string. This method differs from PHP http_build_query as
-     * it does not modify parameters keys.
-     *
-     * @param array<array{0:string, 1:string|float|int|bool|null}> $pairs
-     *
-     * @throws SyntaxError If the encoding type is invalid
-     * @throws SyntaxError If a pair is invalid
-     */
-    public static function build(iterable $pairs, string $separator = '&', int $encType = PHP_QUERY_RFC3986): ?string
-    {
-        $res = match (true) {
-            '' === $separator => throw new SyntaxError('The separator character can not be the empty string.'),
-            !isset(self::ENCODING_LIST[$encType]) => throw new SyntaxError('Unknown or Unsupported encoding.'),
-            default => [],
-        };
-
-        self::$regexpValue = '/(%[A-Fa-f0-9]{2})|[^A-Za-z0-9_\-\.~'.preg_quote(
-            str_replace(
-                html_entity_decode($separator, ENT_HTML5, 'UTF-8'),
-                '',
-                self::ENCODING_LIST[$encType]['suffixValue']
-            ),
-            '/'
-        ).']+/ux';
-
-        self::$regexpKey = '/(%[A-Fa-f0-9]{2})|[^A-Za-z0-9_\-\.~'.preg_quote(
-            str_replace(
-                html_entity_decode($separator, ENT_HTML5, 'UTF-8'),
-                '',
-                self::ENCODING_LIST[$encType]['suffixKey']
-            ),
-            '/'
-        ).']+/ux';
-
-        foreach ($pairs as $pair) {
-            $res[] = self::buildPair($pair);
-        }
-
-        return match (true) {
-            [] === $res => null,
-            PHP_QUERY_RFC1738 === $encType => str_replace('%20', '+', implode($separator, $res)),
-            default => implode($separator, $res),
-        };
-    }
-
-    /**
-     * Build a RFC3986 query key/value pair association.
-     *
-     * @throws SyntaxError If the pair is invalid
-     */
-    private static function buildPair(array $pair): string
-    {
-        if ([0, 1] !== array_keys($pair)) {
-            throw new SyntaxError('A pair must be a sequential array starting at `0` and containing two elements.');
-        }
-
-        [$name, $value] = $pair;
-
-        $encodeMatches = static fn (array $matches): string => match (true) {
-            1 === preg_match(self::REGEXP_UNRESERVED_CHAR, rawurldecode($matches[0])) => rawurlencode($matches[0]),
-            default => $matches[0],
-        };
-
-        $formatter = static fn (string $value, string $regexp): string => match (true) {
-            1 === preg_match('/[\x00-\x1f\x7f]/', $value) => rawurlencode($value),
-            1 === preg_match($regexp, $value) => (string) preg_replace_callback($regexp, $encodeMatches(...), $value),
-            default => $value,
-        };
-
-        $name = match (true) {
-            !is_scalar($name) => throw new SyntaxError(sprintf('A pair key must be a scalar value `%s` given.', gettype($name))),
-            is_bool($name) => (int) $name,
-            is_string($name) => $formatter($name, self::$regexpKey),
-            default => $name,
-        };
-
-        return match (true) {
-            is_string($value) => $name.'='.$formatter($value, self::$regexpValue),
-            is_numeric($value) => $name.'='.$value,
-            is_bool($value) => $name.'='.(int) $value,
-            null === $value => (string) $name,
-            default => throw new SyntaxError(sprintf('A pair value must be a scalar value or the null value, `%s` given.', gettype($value))),
-        };
     }
 }